refactor: 재생 관련 로직 store로 옮김
