+++ src/app/playlist/[id]/page.tsx
@@ -5,7 +5,6 @@ import { useEffect, useState, KeyboardEvent } from "react";
 import styles from "./PlaylistDetail.module.css";
 import { fetchPlaylistTracks, removeTrackFromPlaylist, updatePlaylistName } from "@/apis/userApi";
 import { useSession } from "next-auth/react";
-import { usePlayControl } from "@/hooks/usePlayTracks";
 import { useUIStore } from "@/store/useUiStore";
 import { FaPlay, FaRegEdit, FaCheck, FaTimes, FaRegTrashAlt } from "react-icons/fa";
 import LoadingDots from "@/components/loading/LoadingDots/LoadingDots";
@@ -16,6 +15,7 @@ import { uiToast } from "@/lib/toasts";
 import Image from "next/image";
 import Link from "next/link";
 import axios from "axios";
+import { usePlayerStore } from "@/store/usePlayerStore";
 
 export default function PlaylistDetailPage() {
   const { data: session } = useSession();
@@ -25,6 +25,8 @@ export default function PlaylistDetailPage() {
   const searchParams = useSearchParams();
 
   const { isRelaxMode } = useUIStore();
+  const playFromPlaylist = usePlayerStore((state) => state.playFromPlaylist);
+
   const [tracks, setTracks] = useState<Track[]>([]);
   const [loading, setLoading] = useState(true);
 
@@ -36,8 +38,6 @@ export default function PlaylistDetailPage() {
 
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [selectedTrackUri, setSelectedTrackUri] = useState<string | null>(null);
-
-  const { playFromPlaylist } = usePlayControl();
 
   useEffect(() => {
     if (!token || !id) return;
