+++ src/store/usePlayerStore.tsx
@@ -1,6 +1,9 @@
-import { setRepeatMode, setShuffle } from "@/apis/playbackApi";
+import { setRepeatMode, setShuffle, startPlayback } from "@/apis/playbackApi";
 import { mapSdkTrackToLocalTrack } from "@/lib/spotifyMapper";
+import { uiToast } from "@/lib/toasts";
 import { PlayerState, RepeatMode } from "@/types/player";
+import axios from "axios";
+import { FiExternalLink, FiLock, FiWifiOff } from "react-icons/fi";
 import { create } from "zustand";
 
 export const usePlayerStore = create<PlayerState>((set, get) => ({
@@ -198,6 +201,104 @@ export const usePlayerStore = create<PlayerState>((set, get) => ({
     } catch (e) {
       console.error("반복 모드 변경 실패:", e);
       set({ repeatMode });
+    }
+  },
+
+  playFromPlaylist: async (tracks, startIndex, token) => {
+    // 1. get()을 통해 스토어 내부의 현재 상태들을 모두 가져옴
+    const {
+      deviceId,
+      queue,
+      currentIndex,
+      isPlaying: wasPlaying,
+      optimisticPlay,
+      setIsPlaying,
+      setPosition,
+      setDuration,
+    } = get();
+
+    if (!deviceId || !token) return;
+
+    // 2. 에러 시 롤백을 위한 상태 백업 (컴포넌트가 아니라 스토어 내부에서 안전하게 저장)
+    const previousQueue = queue;
+    const previousIndex = currentIndex;
+    const previousPlaying = wasPlaying;
+
+    // 3. 낙관적 업데이트 실행
+    optimisticPlay(tracks, startIndex);
+    const uris = tracks.map((t) => `spotify:track:${t.id}`);
+
+    try {
+      // 4. 실제 API 호출
+      await startPlayback(uris, deviceId, token, startIndex);
+    } catch (error) {
+      console.error("재생 요청 실패:", error);
+
+      if (axios.isAxiosError(error)) {
+        const status = error.response?.status;
+        const code = error.code;
+
+        // --- 에러 UI 처리 (토스트) ---
+        if (status === 403) {
+          const authLink = (
+            <a
+              href="https://open.spotify.com"
+              target="_blank"
+              rel="noopener noreferrer"
+              className="icon-btn"
+              title="스포티파이 웹에서 인증하기"
+            >
+              <FiExternalLink size="1.4rem" />
+            </a>
+          );
+          uiToast.action(
+            "재생 불가 (성인 인증 필요)",
+            <FiLock
+              size="1.6rem"
+              color="#ff5555"
+            />,
+            authLink,
+            "403-error",
+          );
+        } else if (status === 404) {
+          uiToast.error("플레이어가 비활성화되었습니다. 새로고침 해주세요.", "device-404");
+        } else if (code === "ERR_NETWORK" || error.message === "Network Error") {
+          uiToast.custom(
+            "네트워크 연결이 불안정합니다.",
+            <FiWifiOff
+              size="1.6rem"
+              color="#ff5555"
+            />,
+            "net-error",
+          );
+          setIsPlaying(false);
+          setPosition(0);
+          setDuration(0);
+        } else {
+          uiToast.error(
+            `일시적인 오류가 발생했습니다. ${status ? status : null}`,
+            `error-${status}`,
+          );
+        }
+
+        // --- 상태 롤백 처리 ---
+        if (status === 403 || status === 404) {
+          if (previousQueue.length > 0) {
+            // 아까 백업해둔 상태로 스토어 자체에서 복구
+            optimisticPlay(previousQueue, previousIndex);
+            setIsPlaying(previousPlaying);
+          } else {
+            setIsPlaying(false);
+            setPosition(0);
+            setDuration(0);
+          }
+        } else {
+          setIsPlaying(false);
+        }
+      } else {
+        uiToast.error("알 수 없는 오류가 발생했습니다. 새로고침 해주세요.");
+        setIsPlaying(false);
+      }
     }
   },
 }));
